requirements.txt
django==6.0.3
djangorestframework==3.16.1
django-cors-headers==4.9.0
channels==4.2.0
daphne==4.1.2
python-dotenv==1.0.1
psycopg2-binary==2.9.11
Pillow>=10.0.0

